SDI-259 Moved duplicate helpers to single file
