+++ src/test/java/uk/gov/justice/hmpps/prison/util/builders/BuilderHelpers.kt
@@ -1,8 +1,26 @@
 package uk.gov.justice.hmpps.prison.util.builders
 
+import org.springframework.http.HttpHeaders
+import org.springframework.http.MediaType
 import org.springframework.test.web.reactive.server.WebTestClient
+import org.springframework.test.web.reactive.server.returnResult
+import org.springframework.web.reactive.function.BodyInserters
+import uk.gov.justice.hmpps.prison.api.model.CaseNote
+import uk.gov.justice.hmpps.prison.api.model.CourtHearing
+import uk.gov.justice.hmpps.prison.api.model.PrivilegeSummary
+import uk.gov.justice.hmpps.prison.api.model.RequestToTransferOutToCourt
+import uk.gov.justice.hmpps.prison.api.model.VisitBalances
+import uk.gov.justice.hmpps.prison.api.resource.impl.RestResponsePage
+import uk.gov.justice.hmpps.prison.repository.jpa.model.BedAssignmentHistory
+import uk.gov.justice.hmpps.prison.repository.jpa.model.CourtEvent
+import uk.gov.justice.hmpps.prison.repository.jpa.model.ExternalMovement
 import uk.gov.justice.hmpps.prison.service.DataLoaderRepository
 import uk.gov.justice.hmpps.prison.util.JwtAuthenticationHelper
+import uk.gov.justice.hmpps.prison.util.JwtParameters
+import java.time.Duration
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.util.function.Consumer
 
 internal fun randomName(): String {
   // return random name between 3 and 10 characters long
@@ -16,3 +34,183 @@ data class BuilderContext(
   val jwtAuthenticationHelper: JwtAuthenticationHelper,
   val dataLoader: DataLoaderRepository
 )
+
+fun BuilderContext.transferOut(offenderNo: String, toLocation: String = "MDI") {
+  webTestClient.put()
+    .uri("/api/offenders/{nomsId}/transfer-out", offenderNo)
+    .headers(setAuthorisation(listOf("ROLE_TRANSFER_PRISONER")))
+    .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+    .accept(MediaType.APPLICATION_JSON)
+    .body(
+      BodyInserters.fromValue(
+        """
+          {
+            "transferReasonCode":"NOTR",
+            "commentText":"transferred prisoner today",
+            "toLocation":"$toLocation",
+            "movementTime": "${LocalDateTime.now().minusHours(1).format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)}"
+            
+          }
+        """.trimIndent()
+      )
+    )
+    .exchange()
+    .expectStatus().isOk
+    .expectBody()
+    .jsonPath("inOutStatus").isEqualTo("TRN")
+    .jsonPath("status").isEqualTo("INACTIVE TRN")
+    .jsonPath("lastMovementTypeCode").isEqualTo("TRN")
+    .jsonPath("lastMovementReasonCode").isEqualTo("NOTR")
+    .jsonPath("assignedLivingUnit.agencyId").isEqualTo("TRN")
+    .jsonPath("assignedLivingUnit.description").doesNotExist()
+}
+
+fun BuilderContext.release(offenderNo: String) {
+  webTestClient.put()
+    .uri("/api/offenders/{nomsId}/release", offenderNo)
+    .headers(setAuthorisation(listOf("ROLE_RELEASE_PRISONER")))
+    .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+    .accept(MediaType.APPLICATION_JSON)
+    .body(
+      BodyInserters.fromValue(
+        """
+          {
+            "movementReasonCode":"CR",
+            "commentText":"released prisoner today",
+            "movementTime": "${LocalDateTime.now().minusHours(1).format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)}"
+            
+          }
+        """.trimIndent()
+      )
+    )
+    .exchange()
+    .expectStatus().isOk
+    .expectBody()
+    .jsonPath("inOutStatus").isEqualTo("OUT")
+    .jsonPath("status").isEqualTo("INACTIVE OUT")
+    .jsonPath("lastMovementTypeCode").isEqualTo("REL")
+    .jsonPath("lastMovementReasonCode").isEqualTo("CR")
+    .jsonPath("assignedLivingUnit.agencyId").isEqualTo("OUT")
+    .jsonPath("assignedLivingUnit.description").doesNotExist()
+}
+
+fun BuilderContext.transferOutToCourt(offenderNo: String, toLocation: String, shouldReleaseBed: Boolean = false, courtHearingEventId: Long? = null): LocalDateTime {
+  val movementTime = LocalDateTime.now().minusHours(1)
+  val request = RequestToTransferOutToCourt(
+    /* toLocation = */ toLocation,
+    /* movementTime = */ movementTime,
+    /* escortType = */ null,
+    /* transferReasonCode = */ "19",
+    /* commentText = */ "court appearance",
+    /* shouldReleaseBed = */ shouldReleaseBed,
+    /* courtEventId = */ courtHearingEventId
+  )
+  webTestClient.put()
+    .uri("/api/offenders/{nomsId}/court-transfer-out", offenderNo)
+    .headers(setAuthorisation(listOf("ROLE_TRANSFER_PRISONER_ALPHA")))
+    .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+    .accept(MediaType.APPLICATION_JSON)
+    .body(BodyInserters.fromValue(request))
+    .exchange()
+    .expectStatus().isOk
+    .expectBody()
+    .jsonPath("inOutStatus").isEqualTo("OUT")
+    .jsonPath("status").isEqualTo("ACTIVE OUT")
+    .jsonPath("lastMovementTypeCode").isEqualTo("CRT")
+    .jsonPath("lastMovementReasonCode").isEqualTo("19")
+    .jsonPath("assignedLivingUnit.agencyId").isEqualTo("LEI")
+
+  return movementTime
+}
+
+private fun BuilderContext.setAuthorisation(roles: List<String>): Consumer<HttpHeaders> {
+  return Consumer { httpHeaders: HttpHeaders ->
+    httpHeaders.add(
+      "Authorization",
+      "Bearer " + this.validToken(roles)
+    )
+  }
+}
+
+fun BuilderContext.validToken(roles: List<String?>?): String? {
+  return this.jwtAuthenticationHelper.createJwt(
+    JwtParameters.builder()
+      .username("ITAG_USER")
+      .scope(listOf("read", "write"))
+      .roles(roles)
+      .expiryTime(Duration.ofDays((365 * 10).toLong()))
+      .build()
+  )
+}
+
+fun BuilderContext.createCourtHearing(bookingId: Long): Long {
+  val courtCase = this.dataLoader.offenderCourtCaseRepository.findAllByOffenderBooking_BookingId(bookingId).first()
+
+  return webTestClient.post()
+    .uri("/api/bookings/{bookingId}/court-cases/{courtCaseId}/prison-to-court-hearings", bookingId, courtCase.id)
+    .headers(setAuthorisation(listOf("ROLE_COURT_HEARING_MAINTAINER")))
+    .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+    .accept(MediaType.APPLICATION_JSON)
+    .body(
+      BodyInserters.fromValue(
+        """
+          {
+            "fromPrisonLocation":"LEI",
+            "toCourtLocation":"COURT1",
+            "courtHearingDateTime": "${LocalDateTime.now().plusHours(1).format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)}",
+            "comments":"court appearance"
+            
+          }
+        """.trimIndent()
+      )
+    )
+    .exchange()
+    .expectStatus().isCreated
+    .returnResult<CourtHearing>().responseBody.blockFirst()?.id!!
+}
+
+fun BuilderContext.getMovements(bookingId: Long): List<ExternalMovement> = this.dataLoader.externalMovementRepository.findAllByOffenderBooking_BookingId(bookingId)
+fun BuilderContext.getBedAssignments(bookingId: Long): List<BedAssignmentHistory> =
+  this.dataLoader.bedAssignmentHistoriesRepository.findAllByBedAssignmentHistoryPKOffenderBookingId(bookingId)
+
+fun BuilderContext.getCurrentIEP(offenderNo: String) = webTestClient.get()
+  .uri("/api/offenders/{offenderNo}/iepSummary", offenderNo)
+  .headers(
+    setAuthorisation(
+      listOf("ROLE_SYSTEM_USER")
+    )
+  )
+  .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+  .accept(MediaType.APPLICATION_JSON)
+  .exchange()
+  .expectStatus().isOk
+  .returnResult<PrivilegeSummary>().responseBody.blockFirst()!!
+
+fun BuilderContext.getVOBalanceDetails(offenderNo: String) = webTestClient.get()
+  .uri("/api/bookings/offenderNo/{offenderNo}/visit/balances", offenderNo)
+  .headers(
+    setAuthorisation(
+      listOf("ROLE_SYSTEM_USER")
+    )
+  )
+  .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+  .accept(MediaType.APPLICATION_JSON)
+  .exchange()
+  .expectStatus().isOk
+  .returnResult<VisitBalances>().responseBody.blockFirst()!!
+
+fun BuilderContext.getCaseNotes(bookingId: Long): List<CaseNote> = webTestClient.get()
+  .uri("/api/bookings/{bookingId}/caseNotes?size=999", bookingId)
+  .headers(
+    setAuthorisation(
+      listOf("ROLE_SYSTEM_USER")
+    )
+  )
+  .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+  .accept(MediaType.APPLICATION_JSON)
+  .exchange()
+  .expectStatus().isOk
+  .returnResult<RestResponsePage<CaseNote>>().responseBody.blockFirst()!!.content
+
+fun BuilderContext.getCourtHearings(bookingId: Long): List<CourtEvent> =
+  this.dataLoader.courtEventRepository.findByOffenderBooking_BookingIdOrderByIdAsc(bookingId)
